Add files via upload
This little game i made for my mobile programming exam is based on an elementary idea. You have to order numbers from 1 to 8 starting from the top-left corner.
All you have to do to try it is to upload the project in Android Studio and launch the game with an emulator (you can choose the device to emulate).